Title: ai/report: centralize problem text resolution
- add shared resolver service for problem text
- update AI input and workspace report paths to use the shared resolver
- remove duplicated lookup logic

Files:
- app_server/services/problemText.js
- app_server/services/ai.js
-app_server/datasource/api/workspaceApi.js
